구현위치 변경하고 MainPage에서도 제거했습니다. 그리고 각종 경고문들 있는거 해결했습니다.

+++ 08_IA_project/SpiritManager.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 
 std::string SpiritManager::getSummary() {
-    int count = spirits.size();
+    size_t count = spirits.size();
     double totalYield = 0, totalAbv = 0;
 
     for (const auto& s : spirits) {
